+++ test/test_acoustic_feature_extractor.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from voicevox_engine.acoustic_feature_extractor import OjtPhoneme
+from voicevox_engine.tts_pipeline.acoustic_feature_extractor import OjtPhoneme
 
 TRUE_NUM_PHONEME = 45
