+++ test/test_kana_parser.py
@@ -1,8 +1,8 @@
 from typing import List
 from unittest import TestCase
 
-from voicevox_engine import kana_parser
-from voicevox_engine.kana_parser import create_kana
+from voicevox_engine.tts_pipeline import kana_parser
+from voicevox_engine.tts_pipeline.kana_parser import create_kana
 from voicevox_engine.model import AccentPhrase, Mora, ParseKanaError, ParseKanaErrorCode
