+++ test/test_user_dict_model.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 
 from pydantic import ValidationError
 
-from voicevox_engine.kana_parser import parse_kana
+from voicevox_engine.tts_pipeline.kana_parser import parse_kana
 from voicevox_engine.model import UserDictWord
